Fixed random memory corruption

+++ arch/x86_64/h/defs.h
@@ -1,9 +1,9 @@
 #ifndef defs_h
 #define defs_h
 
-typedef unsigned long virt_addr_t;
-typedef unsigned long virt_size_t;
+typedef unsigned long long virt_addr_t;
+typedef unsigned long long virt_size_t;
 
-typedef unsigned long phys_addr_t;
-typedef unsigned long phys_size_t;
+typedef unsigned long long phys_addr_t;
+typedef unsigned long long phys_size_t;
 #endif
